+++ NEWS.md
@@ -44,9 +44,9 @@ Version 0.16b0 - dev
 - NEW: `ezdxf.math.clip_polygon_2d()`, clip polygon by a convex clipping polygon 
 - NEW: `ezdxf.math.basic_transformation()`, returns a combined transformation
   matrix for translation, scaling and rotation about the z-axis
-- CHANGED: `fit_points_to_cad_cv()` creates a visual equal SPLINE from fit points 
-  like BricsCAD and AutoCAD, but defined by control vertices for all given 
-  scenarios. Removed unused arguments degree and method.   
+- NEW: `fit_points_to_cubic_bezier()` creates a visual equal SPLINE from fit 
+  points without end tangents like BricsCAD, but only for short B-splines.
+- CHANGED: `fit_points_to_cad_cv()`, removed unused arguments `degree` and `method`   
 - CHANGED: `ezdxf.render.nesting` content moved into the `ezdxf.path` package
 - CHANGED: renamed `MeshBuilder.render()` to `MeshBuilder.render_mesh()`
 - DEPRECATED: `ezdxf.render.path` module, replaced by `ezdxf.path` package